fix переплата с прошлого месяца

+++ table_tariff.h
@@ -11,7 +11,7 @@ public:
     table_tariff();
     ~table_tariff();
 
-    double is_Tariff(int usluga, QDate date, int tarif = 1);           //возвращает размер тарифа
+    float is_Tariff(int usluga, QDate date, int tarif = 1);           //возвращает размер тарифа
     QSqlQueryModel* ModelTarifTabl(int year,int month);                         //тарифы на месяц
     void UpdateTarif(QString tarif,QString tarif2, QString norma, int idusl);
     int FillTarif(int month, int year);
